requirements.txt
torch
diffusers
transformers
accelerate
peft
gradio
opencv-python
numpy
pillow
